requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
numpy>=1.24.0
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
fpdf2>=2.7.0
kaggle>=1.5.16
python-dotenv>=1.0.0
